More unclusterfuckering.

+++ src/server/game/Server/WorldSocket.h
@@ -129,7 +129,7 @@ class WorldSocket : protected WorldHandler
         virtual int open (void *);
 
         // Called on failures inside of the acceptor, don't call from your code.
-        virtual int close(u_long);
+        virtual int close(int);
 
         // Called when we can read from the socket.
         virtual int handle_input (ACE_HANDLE = ACE_INVALID_HANDLE);
